requirements.txt
uvicorn==0.21.1
sqlalchemy==1.4.41
fastapi==0.93.0
fastapi-utils==0.2.1
pydantic==1.10.7
passlib==1.7.4
sqlmodel==0.0.8
openai==0.27.0
python-dotenv==0.21.0
python-jose==3.3.0
sqlalchemy-utils==0.41.1
psycopg2
langchain
ptvsd